Moving internal functions to Provider class
